12.2 repair normal views changesAgoDeys(Исправили правильное отображение сколько заказ без изменения)

+++ src/main/java/com/hunt/otziv/p_products/dto/OrderDTOList.java
@@ -61,7 +61,7 @@ public class OrderDTOList {
 
     private LocalDate payDay;
 
-    private int dayToChangeStatusAgo;
+    private long dayToChangeStatusAgo;
